fix(tt): tuning flow end to end
