+++ apps/total-typescript/.env.development
@@ -25,3 +25,5 @@ NEXT_PUBLIC_DISCORD_INVITE_URL=https://discord.gg/8S5ujhfTB3
 EMAIL_SERVER_HOST=smtp.postmarkapp.com
 EMAIL_SERVER_PORT=587
 LOG_VERIFICATION_URL=true
+
+CONVERTKIT_PURCHASED_TAG_ID=3452364
