+++ apps/total-typescript/.env.production
@@ -21,3 +21,4 @@ NEXT_PUBLIC_AMPLITUDE_API_KEY=521aa37b0f147ae1fdf1ad7be34ff9c3
 NEXT_PUBLIC_DISCORD_INVITE_URL=https://discord.gg/8S5ujhfTB3
 EMAIL_SERVER_HOST=smtp.postmarkapp.com
 EMAIL_SERVER_PORT=587
+CONVERTKIT_PURCHASED_TAG_ID=3452364
