+++ packages/skill-api/src/server/index.ts
@@ -3,4 +3,3 @@ export * from './skill-next-auth-prisma-adapter'
 export * from './send-verification-request'
 export * from './default-next-auth-options'
 export * from './send-server-email'
-export * from './get-login-url'
