+++ packages/skill-api/src/server/send-server-email.ts
@@ -1,7 +1,20 @@
 import {NextAuthOptions} from 'next-auth'
+import {createHash, randomBytes} from 'crypto'
 import type {MagicLinkEmailType} from './send-verification-request'
 import {sendVerificationRequest} from './send-verification-request'
-import {baseUrl, getLoginUrl} from './get-login-url'
+import {v4} from 'uuid'
+
+const baseUrl = process.env.NEXTAUTH_URL || process.env.VERCEL_URL
+
+function hashToken(token: string, options: any) {
+  const {provider, secret} = options
+  return (
+    createHash('sha256')
+      // Prefer provider specific secret, but use default secret if none specified
+      .update(`${token}${provider.secret ?? secret}`)
+      .digest('hex')
+  )
+}
 
 export async function sendServerEmail({
   email,
@@ -16,16 +29,36 @@ export async function sendServerEmail({
 }) {
   if (!nextAuthOptions) return
 
+  const emailProvider: any = nextAuthOptions.providers.find(
+    (provider) => provider.id === 'email',
+  )
+
   callbackUrl = (callbackUrl || baseUrl) as string
-  const {token, url, emailProvider, expires} = await getLoginUrl({
-    callbackUrl,
-    identifier: email,
-    nextAuthOptions,
+
+  const identifier = email
+
+  const token = (await emailProvider.generateVerificationToken?.()) ?? v4()
+
+  const ONE_DAY_IN_SECONDS = 86400
+  const expires = new Date(
+    Date.now() + (emailProvider.maxAge ?? ONE_DAY_IN_SECONDS) * 1000,
+  )
+
+  await nextAuthOptions?.adapter?.createVerificationToken?.({
+    identifier,
+    token: hashToken(token, {
+      provider: emailProvider,
+      secret: nextAuthOptions.secret,
+    }),
+    expires,
   })
 
+  const params = new URLSearchParams({callbackUrl, token, email: identifier})
+  const _url = `${baseUrl}/api/auth/callback/${emailProvider.id}?${params}`
+
   await sendVerificationRequest({
-    identifier: email,
-    url,
+    identifier,
+    url: _url,
     theme: nextAuthOptions.theme || {colorScheme: 'auto'},
     provider: emailProvider.options,
     token: token as string,
